+++ blokus_django/chat/urls.py
@@ -1,8 +1,4 @@
 from django.urls import path
-from .views import register, login, lobby
+from . import views
 
-urlpatterns = [
-    path('register/', register, name='register'),
-    path('login/', login, name='login'),
-    path("", lobby)
-]
+urlpatterns = [path("", views.lobby)]
