+++ blokus_django/chat/views.py
@@ -1,27 +1,4 @@
 from django.shortcuts import render
-from django.contrib.auth import authenticate
-from rest_framework.authtoken.models import Token
-from rest_framework.decorators import api_view
-from rest_framework.response import Response
-from blokus_django.serializers import UserSerializer
-
-@api_view(['POST'])
-def register(request):
-    serializer = UserSerializer(data=request.data)
-    if serializer.is_valid():
-        serializer.save()
-        return Response(serializer.data, status=201)
-    return Response(serializer.errors, status=400)
-
-@api_view(['POST'])
-def login(request):
-    username = request.data.get('username')
-    password = request.data.get('password')
-    user = authenticate(username=username, password=password)
-    if user:
-        token, _ = Token.objects.get_or_create(user=user)
-        return Response({'token': token.key})
-    return Response({'error': 'Invalid Credentials'}, status=400)
 
 def lobby(request):
     return render(request, "chat/lobby.html")
