+++ blokus_django/login/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+from .views import register, login
+
+urlpatterns = [
+    path('register/', register, name='register'),
+    path('login/', login, name='login')
+]
