+++ blokus_react/src/App.js
@@ -9,7 +9,7 @@ export class App extends React.Component {
   render() {
     return (
       <div className="App" style={{ width: '100%', height: '100vh' }}>
-        <GamePage  />
+        <Login  />
       </div>
     )
   }
