+++ blokus_react/src/components/login/login.js
@@ -1,11 +1,11 @@
 import React, { useState } from 'react';
-import '../../styles/login/login.css'
+import '../../styles/login/login.css';
+import axios from 'axios';
 
 function Login() {
   const [credentials, setCredentials] = useState({
     username: '',
-    password: '',
-    sessionId: ''
+    password: ''
   });
 
   const handleChange = (e) => {
@@ -16,10 +16,18 @@ function Login() {
     }));
   };
 
-  const handleSubmit = (e) => {
+  const handleSubmit = async (e) => {
     e.preventDefault();
-	//blabla
-    console.log('Anmeldedaten:', credentials);
+    try {
+      const response = await axios.post('http://localhost:8000', { //change later
+        username: credentials.username,
+        password: credentials.password
+      });
+      console.log('Login successful:', response.data);
+      localStorage.setItem('token', response.data.token);
+    } catch (error) {
+      console.error('Login error:', error.response ? error.response.data : error.message);
+    }
   };
 
   return (
@@ -31,10 +39,7 @@ function Login() {
             <input className="login-input" type="text" id="username" name="username" placeholder="USERNAME" value={credentials.username} onChange={handleChange} required />
           </div>
           <div className="input-group">
-            <input  className="login-input" type="password" id="password" name="password" placeholder="PASSWORD" value={credentials.password} onChange={handleChange} required />
-          </div>
-          <div className="input-group">
-            <input  className="login-input" type="text" id="sessionId" name="sessionId" placeholder="SESSIONID" value={credentials.sessionId} onChange={handleChange} required />
+            <input className="login-input" type="password" id="password" name="password" placeholder="PASSWORD" value={credentials.password} onChange={handleChange} required />
           </div>
           <button className="submit-button" type="submit">START</button>
         </form>
